add string filter without ignore case
