+++ core-service/src/main/java/ru/gb/springShop/core/services/soap/GetAllProductSOAPResponse.java
@@ -1,4 +1,4 @@
-//
+package ru.gb.springShop.core.services.soap;//
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
@@ -6,7 +6,6 @@
 //
 
 
-package ru.gb.springShop.core.soap.products;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -38,13 +37,13 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "products"
+    "productsSOAP"
 })
 @XmlRootElement(name = "getAllProductSOAPResponse")
 public class GetAllProductSOAPResponse {
 
     @XmlElement(required = true)
-    protected List<ProductSOAP> products;
+    protected List<ProductSOAP> productsSOAP;
 
     /**
      * Gets the value of the products property.
@@ -69,10 +68,10 @@ public class GetAllProductSOAPResponse {
      * 
      */
     public List<ProductSOAP> getProducts() {
-        if (products == null) {
-            products = new ArrayList<ProductSOAP>();
+        if (productsSOAP == null) {
+            productsSOAP = new ArrayList<ProductSOAP>();
         }
-        return this.products;
+        return this.productsSOAP;
     }
 
 }
