+++ core-service/src/main/java/ru/gb/springShop/core/services/soap/GetProductSOAPByIdRequest.java
@@ -1,4 +1,4 @@
-//
+package ru.gb.springShop.core.services.soap;//
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
@@ -6,11 +6,8 @@
 //
 
 
-package ru.gb.springShop.core.soap.products;
-
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
 
@@ -25,7 +22,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="productSOAP" type="{http://www.flamexander.com/spring/shop/products}productSOAP"/&gt;
+ *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -36,36 +33,27 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "productSOAP"
+    "id"
 })
-@XmlRootElement(name = "getProductSOAPByIdResponse")
-public class GetProductSOAPByIdResponse {
+@XmlRootElement(name = "getProductSOAPByIdRequest")
+public class GetProductSOAPByIdRequest {
 
-    @XmlElement(required = true)
-    protected ProductSOAP productSOAP;
+    protected long id;
 
     /**
-     * Gets the value of the productSOAP property.
+     * Gets the value of the id property.
      * 
-     * @return
-     *     possible object is
-     *     {@link ProductSOAP }
-     *     
      */
-    public ProductSOAP getProductSOAP() {
-        return productSOAP;
+    public long getId() {
+        return id;
     }
 
     /**
-     * Sets the value of the productSOAP property.
+     * Sets the value of the id property.
      * 
-     * @param value
-     *     allowed object is
-     *     {@link ProductSOAP }
-     *     
      */
-    public void setProductSOAP(ProductSOAP value) {
-        this.productSOAP = value;
+    public void setId(long value) {
+        this.id = value;
     }
 
 }
