+++ core-service/src/main/java/ru/gb/springShop/core/services/soap/GetProductSOAPByIdResponse.java
@@ -1,4 +1,4 @@
-//
+package ru.gb.springShop.core.services.soap;//
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
@@ -6,10 +6,10 @@
 //
 
 
-package ru.gb.springShop.core.soap.products;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
 
@@ -24,7 +24,7 @@ import javax.xml.bind.annotation.XmlType;
  *   &lt;complexContent&gt;
  *     &lt;restriction base="{http://www.w3.org/2001/XMLSchema}anyType"&gt;
  *       &lt;sequence&gt;
- *         &lt;element name="id" type="{http://www.w3.org/2001/XMLSchema}long"/&gt;
+ *         &lt;element name="productSOAP" type="{http://www.flamexander.com/spring/shop/products}productSOAP"/&gt;
  *       &lt;/sequence&gt;
  *     &lt;/restriction&gt;
  *   &lt;/complexContent&gt;
@@ -35,27 +35,36 @@ import javax.xml.bind.annotation.XmlType;
  */
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlType(name = "", propOrder = {
-    "id"
+    "productSOAP"
 })
-@XmlRootElement(name = "getProductSOAPByIdRequest")
-public class GetProductSOAPByIdRequest {
+@XmlRootElement(name = "getProductSOAPByIdResponse")
+public class GetProductSOAPByIdResponse {
 
-    protected long id;
+    @XmlElement(required = true)
+    protected ProductSOAP productSOAP;
 
     /**
-     * Gets the value of the id property.
+     * Gets the value of the productSOAP property.
      * 
+     * @return
+     *     possible object is
+     *     {@link ProductSOAP }
+     *     
      */
-    public long getId() {
-        return id;
+    public ProductSOAP getProductSOAP() {
+        return productSOAP;
     }
 
     /**
-     * Sets the value of the id property.
+     * Sets the value of the productSOAP property.
      * 
+     * @param value
+     *     allowed object is
+     *     {@link ProductSOAP }
+     *     
      */
-    public void setId(long value) {
-        this.id = value;
+    public void setProductSOAP(ProductSOAP value) {
+        this.productSOAP = value;
     }
 
 }
