+++ core-service/src/main/java/ru/gb/springShop/core/services/soap/ObjectFactory.java
@@ -1,4 +1,4 @@
-//
+package ru.gb.springShop.core.services.soap;//
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
@@ -6,7 +6,7 @@
 //
 
 
-package ru.gb.springShop.core.soap.products;
+
 
 import javax.xml.bind.annotation.XmlRegistry;
