+++ core-service/src/main/java/ru/gb/springShop/core/services/soap/ProductSOAP.java
@@ -1,4 +1,4 @@
-//
+package ru.gb.springShop.core.services.soap;//
 // This file was generated by the JavaTM Architecture for XML Binding(JAXB) Reference Implementation, v2.3.2 
 // See <a href="https://javaee.github.io/jaxb-v2/">https://javaee.github.io/jaxb-v2/</a> 
 // Any modifications to this file will be lost upon recompilation of the source schema. 
@@ -6,7 +6,6 @@
 //
 
 
-package ru.gb.springShop.core.soap.products;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
@@ -97,7 +96,7 @@ public class ProductSOAP {
      * Gets the value of the price property.
      * 
      */
-    public BigDecimal  getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -105,7 +104,7 @@ public class ProductSOAP {
      * Sets the value of the price property.
      * 
      */
-    public void setPrice(BigDecimal  value) {
+    public void setPrice(BigDecimal value) {
         this.price = value;
     }
