+++ core-service/src/main/resources/productsSOAP.xsd
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <xs:schema
         xmlns:xs="http://www.w3.org/2001/XMLSchema"
-         xmlns:tns="http://www.flamexander.com/spring/soap/products"
-        targetNamespace="http://www.flamexander.com/spring/soap/products" elementFormDefault="qualified">
+         xmlns:tns="http://www.gb.ru/springShop/soap/products"
+        targetNamespace="http://www.gb.ru/springShop/soap/products" elementFormDefault="qualified">
     <!--по id-->
     <xs:element name="getProductSOAPByIdRequest">
         <xs:complexType>
@@ -32,7 +32,7 @@
         <xs:complexType>
             <xs:sequence>
 <!--  ограничени количества (без ограничений)              -->
-                <xs:element name="products" maxOccurs="unbounded" type="tns:productSOAP"/>
+                <xs:element name="productSOAP" maxOccurs="unbounded" type="tns:productSOAP"/>
             </xs:sequence>
         </xs:complexType>
     </xs:element>
